# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DAYANERA Core API",
    "description": "Özel, yalnızca yerel (127.0.0.1) DAYANERA.ai çekirdek API'si. Tarayıcı, gelecekteki Windows ve Android istemcileri bu tek arayüzü kullanır. Herkese açık bir ürün API'si değildir.",
    "version": "0.1.0-beta"
  },
  "paths": {
...
    "/api/v1/memory/search": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Hafıza ve sohbet geçmişinde ara",
        "operationId": "search_api_v1_memory_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 300,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Api V1 Memory Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/items": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Hafıza kayıtları",
        "operationId": "items_api_v1_memory_items_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Items Api V1 Memory Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Hafıza kaydı ekle (açık kullanıcı bilgisi)",
        "operationId": "create_api_v1_memory_items_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Api V1 Memory Items Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/items/{item_id}": {
      "patch": {
        "tags": [
          "memory"
        ],
        "summary": "Hafıza kaydını güncelle (eski kayıt 'superseded' olarak korunur)",
        "operationId": "patch_api_v1_memory_items__item_id__patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryPatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch Api V1 Memory Items  Item Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/items/{item_id}/history": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Kaydın sürüm zinciri",
        "operationId": "history_api_v1_memory_items__item_id__history_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response History Api V1 Memory Items  Item Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notes": {
      "get": {
        "tags": [
          "recommendation-notes"
        ],
        "summary": "Öneri notları",
        "operationId": "list_notes_api_v1_notes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Notes Api V1 Notes Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "recommendation-notes"
        ],
        "summary": "Öneri notu oluştur (yalnızca agent-notes/)",
        "operationId": "create_note_api_v1_notes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Note Api V1 Notes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/{filename}": {
      "get": {
        "tags": [
          "recommendation-notes"
        ],
        "summary": "Öneri notu (Markdown)",
        "operationId": "read_note_api_v1_notes__filename__get",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Read Note Api V1 Notes  Filename  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "recommendation-notes"
        ],
        "summary": "Öneri notunu güncelle (durum/içerik)",
        "operationId": "update_note_api_v1_notes__filename__patch",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotePatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Note Api V1 Notes  Filename  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryIn": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "fact",
              "preference",
              "technical_value",
              "source_link",
              "relationship",
              "note"
            ],
            "title": "Kind",
            "default": "fact"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Content"
          },
          "status": {
            "type": "string",
            "enum": [
              "user_confirmed",
              "unverified"
            ],
            "title": "Status",
            "default": "user_confirmed"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "shared"
            ],
            "title": "Visibility",
            "default": "private"
          },
          "structured": {
            "additionalProperties": true,
            "type": "object",
            "title": "Structured"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "MemoryIn"
      },
      "MemoryPatchIn": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "structured": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "user_confirmed",
                  "unverified",
                  "archived",
                  "deleted"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "MemoryPatchIn"
      },
...
      "NoteIn": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title"
          },
          "recommendation": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Recommendation"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "default": ""
          },
          "affected_areas": {
            "type": "string",
            "title": "Affected Areas",
            "default": ""
          },
          "expected_benefit": {
            "type": "string",
            "title": "Expected Benefit",
            "default": ""
          },
          "risks": {
            "type": "string",
            "title": "Risks",
            "default": ""
          },
          "status": {
            "type": "string",
            "enum": [
              "öneri",
              "inceleniyor",
              "uygulandı",
              "reddedildi"
            ],
            "title": "Status",
            "default": "öneri"
          },
          "author": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          }
        },
        "type": "object",
        "required": [
          "title",
          "recommendation"
        ],
        "title": "NoteIn"
      },
      "NotePatchIn": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "öneri",
                  "inceleniyor",
                  "uygulandı",
                  "reddedildi"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "affected_areas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Areas"
          },
          "expected_benefit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Benefit"
          },
          "risks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risks"
          }
        },
        "type": "object",
        "title": "NotePatchIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}